Move some code into new file to avoid circular deps
