+++ apps/zui/src/views/export-modal/state.ts
@@ -1,7 +1,7 @@
 import {useState} from "react"
 import {useSelector} from "react-redux"
 import Results from "src/js/state/Results"
-import {RESULTS_QUERY_COUNT} from "../results-pane/run-results-query"
+import {RESULTS_QUERY_COUNT} from "../results-pane/config"
 
 export function useExportModalState() {
   const countStatus = useSelector(Results.getStatus(RESULTS_QUERY_COUNT))
