+++ apps/zui/src/views/results-pane/config.ts
@@ -1,0 +1,2 @@
+export const RESULTS_QUERY = "zui-results/main"
+export const RESULTS_QUERY_COUNT = "zui-results/main-count"
